import from SVN
I was not able to work any magic to get all the svn commits, so
I took the latest version of trunk.

Note, I did work a bit with Claude to modernize a few things and
clean things up a bit, improve the testing, and generate better
docs.

This new code is untested at this point, but generally looks
okay, given an eye test.